pyproject.toml
[project]
name = "ada502_fireguard"
version = "0.1.0"
description = "FireGuard"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask",
    "folium>=0.20.0",
    "requests>=2.32.5",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ada502_fireguard"]
